+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 nylas-python Changelog
 ======================
 
-Unreleased
+v6.13.0
 ----------
 * Fixed from field handling in messages.send() to properly map "from_" field to "from field
 * Fixed content_id handling for large inline attachments to use content_id as field name instead of generic file{index}